backend/requirements.txt
sqlalchemy[asyncio] 
asyncpg
pydantic
fastapi
uvicorn
python-dotenv
alembic
langchain
langchain_huggingface
langchain_community
passlib[bcrypt]
python-jose
chromadb
pymupdf
python-multipart
